Added NCU-Core minified, changed filename of NCU-Core.js
The NCU-Core.js didn't need the version number in the file name. It's located in the comments, and will be packaged with the release.

Added a new FormFiller.Events event to handle resizing the form after a formatting rule has been processed wiht the RegisterRuleProcessed event. Should now correctly resize the form in every possible situation so it will not grow / shrink after several resizes.

NCU-Core-min.js is the library compressed, but likely not as efficiently as it could be. We'll get to that later.